+++ shared/java/svg/SVGLengthType.java
@@ -2,6 +2,7 @@ package io.github.humbleui.skija.svg;
 
 import org.jetbrains.annotations.*;
 
+// SkSVGRenderContext.h - SkSVGLengthContext::LengthType
 public enum SVGLengthType {
     HORIZONTAL,
     VERTICAL,