+++ shared/java/svg/SVGPreserveAspectRatioScale.java
@@ -2,6 +2,7 @@ package io.github.humbleui.skija.svg;
 
 import org.jetbrains.annotations.*;
 
+// SkSVGTypes.h - SkSVGPreserveAspectRatio::Scale
 public enum SVGPreserveAspectRatioScale {
     MEET,
     SLICE;