+++ shared/java/svg/SVGTag.java
@@ -2,6 +2,7 @@ package io.github.humbleui.skija.svg;
 
 import org.jetbrains.annotations.*;
 
+// SkSVGNode.h - SkSVGTag
 public enum SVGTag {
     CIRCLE,
     CLIP_PATH,
@@ -9,13 +10,20 @@ public enum SVGTag {
     ELLIPSE,
     FE_BLEND,
     FE_COLOR_MATRIX,
+    FE_COMPONENT_TRANSFER,
     FE_COMPOSITE,
     FE_DIFFUSE_LIGHTING,
     FE_DISPLACEMENT_MAP,
     FE_DISTANT_LIGHT,
     FE_FLOOD,
+    FE_FUNC_A,
+    FE_FUNC_R,
+    FE_FUNC_G,
+    FE_FUNC_B,
     FE_GAUSSIAN_BLUR,
     FE_IMAGE,
+    FE_MERGE,
+    FE_MERGE_NODE,
     FE_MORPHOLOGY,
     FE_OFFSET,
     FE_POINT_LIGHT,